refactor: test시 동적으로 환경변수를 읽어오게 바꾸었다.
